Restructured the Folders and CleanUp Code
